+++ app/Enums/PlatformAuditAction.php
@@ -8,4 +8,8 @@ enum PlatformAuditAction: string
     case AdministratorRevoked = 'administrator.revoked';
     case MemberSuspended = 'member.suspended';
     case MemberReinstated = 'member.reinstated';
+    case DirectMessageReportReviewStarted = 'direct_message_report.reviewing';
+    case DirectMessageReportResolved = 'direct_message_report.resolved';
+    case DirectMessageReportDismissed = 'direct_message_report.dismissed';
+    case DirectMessageReportReopened = 'direct_message_report.reopened';
 }